// Backend/src/config-ranking/types/ranking-config-sync.type.ts
// types/ranking-config-sync.type.ts

import { ModoRanking } from "@prisma/client";

export type RankingConfigSync = {
  id: number;
  modo: ModoRanking;

  fecha_inicio: string;
  fecha_fin: string;
  fecha_corte: string;

  zona_horaria: string;

  hora_inicio_actualizacion: string;
  hora_fin_actualizacion: string;
  intervalo_actualizacion: number;

  ultima_actualizacion: string | null;
  proxima_actualizacion: string | null;

  updateAt: Date;
};